Extract classes from single file
